+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # ChangeLog
 
+- add DuckDB::Appender
 - bump DuckDB to v0.2.6 in CI.
 - remove unnecessary duckdb header file from MacOS CI.
 - add DuckDB::Connection#prepared_statement.